Move all the collections
